deleted gitignore
the one secret we had isn't actually a secret. the users can't do any harm with just the URL, because they'd need the API key to access any of the create/update/delete routes anyway. that one actually is secret and is removed from the other two repos lol - removing this here makes it easy to deploy to netlify just by making the update